UI: Move shameful components into a shame/ directory
Change-type: patch
