+++ lib/ui/lens/Interleaved.tsx
@@ -18,11 +18,12 @@ import {
 import styled from 'styled-components';
 import uuid = require('uuid/v4');
 import { Event as EventCard } from '../components/Event';
-import Icon from '../components/Icon';
 import { analytics, sdk } from '../core';
 import { actionCreators, selectors, StoreState } from '../core/store';
 import { createChannel, getUpdateObjectFromSchema, getViewSchema } from '../services/helpers';
 import { Card, Lens, RendererProps } from '../types';
+
+import Icon from '../shame/Icon';
 
 const Column = styled(Flex)`
 	height: 100%;
