+++ lib/ui/lens/List.tsx
@@ -25,11 +25,12 @@ import {
 } from 'rendition';
 import styled from 'styled-components';
 import { CardCreator } from '../components/CardCreator';
-import Icon from '../components/Icon';
 import { actionCreators } from '../core/store';
 import { createChannel, getUpdateObjectFromSchema, getViewSchema } from '../services/helpers';
 import { Card, Lens, RendererProps, Type } from '../types';
 import SingleCardLens from './SingleCard';
+
+import Icon from '../shame/Icon';
 
 const Column = styled(Flex)`
 	height: 100%;
