+++ lib/ui/lens/Org.tsx
@@ -22,7 +22,6 @@ import { Markdown } from 'rendition/dist/extra/Markdown';
 import { Mermaid } from 'rendition/dist/extra/Mermaid';
 import styled from 'styled-components';
 import { CardActions } from '../components/CardActions';
-import { CloseButton } from '../components/CloseButton';
 import Label from '../components/Label';
 import { Tag } from '../components/Tag';
 import { actionCreators, selectors, StoreState } from '../core/store';
@@ -35,6 +34,8 @@ import {
 import { createLink } from '../services/link';
 import { Card, Lens, RendererProps, Type } from '../types';
 import TimelineLens from './Timeline';
+
+import { CloseButton } from '../shame/CloseButton';
 
 const Column = styled(Flex)`
 	height: 100%;
