+++ lib/ui/lens/Table.tsx
@@ -18,10 +18,11 @@ import {
 } from 'rendition';
 import styled from 'styled-components';
 import { CardCreator } from '../components/CardCreator';
-import Icon from '../components/Icon';
 import { actionCreators } from '../core/store';
 import { createChannel, getUpdateObjectFromSchema, getViewSchema } from '../services/helpers';
 import { Card, Lens, RendererProps, Type } from '../types';
+
+import Icon from '../shame/Icon';
 
 const Column = styled(Flex)`
 	height: 100%;
