+++ lib/ui/lens/Thread.tsx
@@ -19,7 +19,6 @@ import { Markdown } from 'rendition/dist/extra/Markdown';
 import { Mermaid } from 'rendition/dist/extra/Mermaid';
 import styled from 'styled-components';
 import { CardActions } from '../components/CardActions';
-import { CloseButton } from '../components/CloseButton';
 import Label from '../components/Label';
 import { Tag } from '../components/Tag';
 import { actionCreators, selectors, StoreState } from '../core/store';
@@ -31,6 +30,8 @@ import {
 import { getActor } from '../services/store-helpers';
 import { Card, Lens, RendererProps, Type } from '../types';
 import TimelineLens from './Timeline';
+
+import { CloseButton } from '../shame/CloseButton';
 
 const Column = styled(Flex)`
 	height: 100%;
