+++ lib/ui/shame/CloseButton.tsx
