+++ lib/ui/shame/DataRenderer.tsx
